% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{cap_first_letter}
\alias{cap_first_letter}
\title{Make the first letter upper case}
\usage{
cap_first_letter(x)
}
\arguments{
\item{x}{A vector of species names.}
}
\value{
A vector.
}
\description{
Make the first letter to be upper case.
}
