% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{dist_to_df}
\alias{dist_to_df}
\title{Change a distance matrix or a matrix to a data frame}
\usage{
dist_to_df(x)
}
\arguments{
\item{x}{An object of class 'dist' or a matrix.}
}
\value{
A data frame with three columns: Var1, Var2, distance.
}
\description{
Convert a distance class object or a matrix to a data frame.
}
