% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_tree.R
\name{get_n_polytomy}
\alias{get_n_polytomy}
\title{Calculate number of polytomies}
\usage{
get_n_polytomy(tree)
}
\arguments{
\item{tree}{A phylogeny with 'phylo' as class.}
}
\value{
A data frame with number of polytomies as columns.
}
\description{
Calculate the number of polytomies a phylogeny has.
}
\examples{
library(lirrr) 
get_n_polytomy(tree)

}
