% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{get_pd_alpha}
\alias{get_pd_alpha}
\title{Calculate alpha phylogenetic diversity}
\usage{
get_pd_alpha(samp_wide, tree, samp_long, null.model.pd.root = FALSE,
  null.model.phylomeasures = TRUE, null.model.phylocom = FALSE,
  null.type.phylomeasures = "uniform", null.type.phylocom = 0,
  null.type.picante = "taxa.labels", n.item = 999, abund.weight = FALSE,
  verbose = TRUE, vpd = FALSE, ...)
}
\arguments{
\item{samp_wide}{Wide version of samp_long, row.names are sites, colnames are species.}

\item{tree}{A phylogeny with class of 'phylo'.}

\item{samp_long}{A 3-column data frame, site, freq, sp.}

\item{null.model.pd.root}{Whether to run null models for rooted PD?}

\item{null.type.phylomeasures}{If null.model is TRUE, which null model to use for PhyloMeasure?}

\item{null.type.phylocom}{If null.model is TRUE, which null model to use for Phylocom? 
- 0: phylogeny shuffle. 
- 1: maintain site richness and draw from species actually observed in sites. 
- 2: maintain site richness and draw from the phylogeny
- 3: independent swap. 
See ?phylocomr::ph_comstruct for details}

\item{null.type.picante}{If null.model is TRUE, which null model to use for Picante? Not used yet.}

\item{n.item}{The number of randomization.}

\item{abund.weight}{Should abundance information used when calculating pd with Phylocom/Picante? Default is FALSE.}

\item{verbose}{Do you want to see relevant information?}

\item{vpd}{To calculate vpd (varance of pairwise distance) or not?}
}
\value{
A data frame.
}
\description{
A function to calculate a bunch of phylo diversity: Faith's PD, MPD, VPD (variance of pairwise distance), MNTD, PSV/PSE.
Results of MPD/MNTD from PhyloMeasures are equal with those from Phylocom/Picante with abundance.weight = FALSE
Results of PD from PhyloMeasures are unrooted; PD from Phylocom are rooted.
}
