% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{highlight_subset_sp_in_phylogeny}
\alias{highlight_subset_sp_in_phylogeny}
\title{Highlight all branches linking a subset of species in a phylogeny}
\usage{
highlight_subset_sp_in_phylogeny(phy, subset_sp, highlight_color = "red")
}
\arguments{
\item{phy}{The large phylogeny to plot.}

\item{subset_sp}{A vector of species names to be highlighted on the plot of phylogeny.}

\item{highlight_color}{The color to highlight the branches, the default is red.}
}
\value{
A `ggplot2` object.
}
\description{
Highlight all branches linking a subset of species in a phylogeny
}
