% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{logit_tran}
\alias{logit_tran}
\title{Logit transformation}
\usage{
logit_tran(x, add_num = 0.01)
}
\arguments{
\item{x}{A vector of proportions (either in form of 0.20 or 20).}

\item{add_num}{The number to add for 0s, the dafult is 0.01.}
}
\value{
A vector has been logit transformed.
}
\description{
Conduct logit-transformation for proportion data.
}
