% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{mvpd}
\alias{mvpd}
\title{MPD and VPD (mean and variance of pairwise distance)}
\usage{
mvpd(samp, dis, abundance.weighted = FALSE)
}
\arguments{
\item{samp}{A site by species data frame, site names as row names.}

\item{dis}{A distance matrix.}

\item{abundance.weighted}{Whether to weight by species abundance? Default is `FALSE`.}
}
\value{
A data frame with three columns: site, mpd, and vpd.
}
\description{
Calculate MPD and VPD. The distance matrix can be from functional traits or phylogeny.
This function is the same as `picante::mpd` when `abundance.weighted = FALSE`;
it is, however, different when `abundance.weighted = TRUE` because this function
does not include the diagonal values while `picante::mpd` does, which is actually equal
to Rao's Q instead of MPD (de Bello et al, 2016, Oecologia).
}
