% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{panel.cor}
\alias{panel.cor}
\title{\code{panel.cor} to plot absolute value of correlations}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, color.threshold = 0.5,
  ...)
}
\description{
\code{panel.cor} to plot absolute value of correlations
}
