% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{rm_site_noobs}
\alias{rm_site_noobs}
\title{Remove site that has no observations (site by sp matrix)}
\usage{
rm_site_noobs(df)
}
\arguments{
\item{df}{A data frame in wide form, i.e. site by species data frame, with site names as row name.}
}
\value{
A data frame.
}
\description{
Remove site that has no obsrevations of any species.
}
