% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{rm_sp_noobs}
\alias{rm_sp_noobs}
\title{Remove sp that not observed at any site (site by sp matrix)}
\usage{
rm_sp_noobs(df)
}
\arguments{
\item{df}{A data frame in wide form, i.e. site by species data frame, with site names as row name.}
}
\value{
A data frame.
}
\description{
Remove species that not observed in any site.
}
