% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{tnrs_match_names_2}
\alias{tnrs_match_names_2}
\title{Match taxa names with the Open Tree of Life}
\usage{
tnrs_match_names_2(taxa, n_per_req = 20, ...)
}
\arguments{
\item{taxa}{A vector of species names to match.}

\item{n_per_req}{Number of species to match per requery, must be <= 250.}
}
\value{
A data frame.
}
\description{
This function will try to do batch match based on rotl::tnrs_match_names(), 
  which only allows <= 250 species each time.
}
