% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean.R
\name{var_to_rownames}
\alias{var_to_rownames}
\title{Function to add a column as row names, and remove it from columns}
\usage{
var_to_rownames(df, var = "site")
}
\arguments{
\item{df}{A data frame.}

\item{var}{The column name (as character) you want to put as row name.}
}
\value{
A data frame.
}
\description{
Convert a column of a data frame to be its row name.
}
