% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{make_world_grids}
\alias{make_world_grids}
\title{Make grid cells across the lands of world}
\usage{
make_world_grids(cell_size = 1e+05, ...)
}
\arguments{
\item{cell_size}{The size of the cell, default is 100 km by 100 km.}

\item{...}{Additional arguments to be passed into `sf::st_make_grid()`.}
}
\value{
A list of two sf objectives: `grid_sf` is the grid cells of the world; 
`grid_land` only contains the grid cells that intersected with terrestrial lands,
which can be further filtered by continent, or country, etc.
}
\description{
Make grid cells across the lands of world
}
